% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{MASPPRE}
\alias{MASPPRE}
\title{This function is applied to the partly linear additive Cox model, and offers the weight estimators of model averaging and the prediction of survival probability of individual given the time point for right censoring data, where the weight selection criteria employ the inverse probability of censoring
weights and the non-parametric parts of this model are approximated by B-splines method.}
\usage{
MASPPRE(t, sim.data, p, l, h, alpha)
}
\arguments{
\item{t}{numeric, the given time point.}

\item{sim.data}{a data.frame, the column names of the data.frame are U.1, U.2, U.3 and U.4 (Continuous covariate), X (The binary covariate), time (Observation time), status (Censor indicator).}

\item{p}{numeric, the dimension of continuous covariate.}

\item{l}{numeric, the dimension of discrete covariate.}

\item{h}{numeric, the number of splines basis functions.}

\item{alpha}{numeric, The partition ratio of training set and test set}
}
\value{
MASPPRE returns a list containing the following components:

\item{weights}{the estimators of weights.}

\item{SP_train}{the prediction of survival probability in the train sets.}

\item{SP_test}{the prediction of survival probability in the test sets.}
}
\description{
This function is applied to the partly linear additive Cox model, and offers the weight estimators of model averaging and the prediction of survival probability of individual given the time point for right censoring data, where the weight selection criteria employ the inverse probability of censoring
weights and the non-parametric parts of this model are approximated by B-splines method.
}
\examples{
data("sim.data")
h <- 5
p <- 4
l <- 1
alpha <- 0.7
t <- mean(sim.data$time)
## estimate weights and predict the survival probability
# MASPPRE(t, sim.data, p, l, h, alpha)
}
